backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai
pydantic
sqlalchemy
psycopg2-binary
redis